requirement.txt
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==2.1.5
django-allauth==0.38.0
django-rest-auth==0.9.3
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
googlemaps==3.0.2
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
oauthlib==3.0.1
Pillow==5.4.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2018.9
PyYAML==3.13
requests==2.11.1
requests-oauthlib==1.2.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
pymysql==0.9.3
drfpasswordless==1.3.7
exponent_server_sdk==0.2.0
pyrebase==3.0.27
